convert special page main component into small component
